Add files via upload
Demonstration of a Generic HTML form solution for satisfying the FDA and GMP requirements for Annual Product Reviews.

This form was developed for Factory-Talk Technical Services, Ltd. in Bangkok, Thailand - to be hosted on MasterControl eQMS software. As far as I know there was only a hand full of demonstrations of this form - and only 1 or 2 sales.

This form is NOT licensed - nor is it being shared for any other reason than to demonstrate the capabilities of the functions written for its' operation.  ALL RIGHTS RESERVED and belong to Factory-Talk Technical Services.